TestProject/requirements.txt
Django==4.2
djangorestframework~=3.14.0
Pillow==9.2.0
python-decouple==3.4 #conf
drf-yasg #swager
django-cors-headers~=4.0.0
channels~=4.0.0
channels-redis~=4.1.0
daphne~=4.0.0
redis==4.5.3
drf-spectacular~=0.26.2
gunicorn

